adding profile screen
