+++ app/DeerTracker/PhotoGallery.js
@@ -154,8 +154,7 @@ export default class PhotoGallery extends React.Component {
         if (crops.length > 0) {
             return (
                 <View style={{ height: 280 }}>
-                    <Swiper loop={false} dot={<View />} activeDot={<View />}
-                        ref={this.swiper}
+                    <Swiper loop={false} ref={this.swiper}
                         onIndexChanged={(nextIndex) => { this.generateCrop(crops[nextIndex]) }}>
                         {crops.map((crop) => {
                             let max = Math.max(crop.w, crop.h);
