+++ app/DeerTracker/PhotoScreen.js
@@ -23,16 +23,18 @@ export default class PhotoScreen extends React.Component {
     const showCrops = this.props.navigation.getParam('showCrops');
     return (
       <SafeAreaView>
-        <View style={{ height: 80 }}>
-          <View style={{ alignItems: 'center' }}>
-            <Text style={style.t4}>{title}</Text>
-          </View>
+        <View style={{ height: title ? 80 : 40 }}>
+          {title &&
+            <View style={{ alignItems: 'center' }}>
+              <Text style={style.t4}>{title}</Text>
+            </View>
+          }
           <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'center' }}>
             <Image source={require('./assets/images/crosshairs.png')} style={{ width: 25, height: 25 }} />
             <Text style={style.t4}>{subTitle}</Text>
           </View>
         </View>
-        <View style={{ height: screenHeight - 80 - headerHeight - footerHeight }}>
+        <View style={{ height: screenHeight - (title ? 80 : 40) - headerHeight - footerHeight }}>
           <PhotoGrid photos={photos} showCrops={showCrops} onRefresh={() => { this.props.navigation.goBack() }} />
         </View>
       </SafeAreaView>
