+++ app/DeerTracker/ProfileScreen.js
@@ -1,132 +1,75 @@
 import React from 'react';
 import {
-  Alert,
   SafeAreaView,
-  ActivityIndicator,
-  ScrollView,
-  RefreshControl,
   Text,
   View,
-  Image,
-  TouchableOpacity,
+  Image
 } from 'react-native';
 
-import Icon from 'react-native-vector-icons/FontAwesome5';
 import RNFS from 'react-native-fs';
-import Moment from 'moment';
 
-import SwipeRow from './SwipeRow';
+import Swiper from 'react-native-swiper'
+import PhotoGrid from './PhotoGrid';
+
+import MoonPhase from './MoonPhase';
 import Database from './Database';
 
-import style, { screenWidth, screenHeight, thumbWidth, thumbHeight, headerHeight } from './style';
+import style, { screenHeight, screenWidth, thumbWidth, headerHeight, footerHeight } from './style';
 
 const root = RNFS.DocumentDirectoryPath;
+
+const moon = new MoonPhase();
 
 export default class ProfileScreen extends React.Component {
 
   constructor(props) {
     super(props);
     this.db = new Database();
-    this.state = { isLoading: true }
-  }
-
-  componentDidMount() {
-    this.fetchData();
-  }
-
-  refreshing() {
-    return this.state.isLoading;
   }
 
   render() {
+    const profile = this.props.navigation.getParam('profile');
+    let photos = profile.objects.map((photo) => {
+      photo.photo_path = root + '/' + photo.photo_path;
+      photo.url = photo.photo_path;
+      photo.props = {
+        photo: photo
+      };
+      photo.objects = [photo];
+      return photo;
+    });
 
-    if (this.refreshing()) {
-      return (
-        <SafeAreaView>
-          <View style={style.activity}>
-            <ActivityIndicator size='large' />
-          </View>
-        </SafeAreaView>
-      )
-    }
-
-    const profiles = this.state.profiles;
+    let crop = photos[0];
+    let max = Math.max(crop.w, crop.h);
+    let w = (screenWidth / max) * crop.w;
+    let h = (140 / max) * crop.h;
 
     return (
       <SafeAreaView>
-        <ScrollView style={{ height: '100%' }} refreshControl={
-          <RefreshControl
-            title='Refresh'
-            titleColor='black'
-            tintColor='black'
-            refreshing={this.refreshing()} onRefresh={this.fetchData.bind(this)} />
-        }>
-          {profiles.length <= 0 &&
-            <Text style={style.t3}>Review sightings to add a profile</Text>
-          }
-          {profiles.length > 0 &&
-            <Text style={style.t3}>Select a profile</Text>
-          }
-          {profiles.map((profile) => {
-            let crop = profile.objects[0];
-            crop.path = RNFS.CachesDirectoryPath + '/crop_' + crop.id + '.jpg';
-            let max = Math.max(crop.w, crop.h);
-            let w = (thumbWidth / max) * crop.w;
-            let h = (140 / max) * crop.h;
-            return (
-              <SwipeRow key={crop.profile_name} item={crop} onDelete={this.deleteProfile.bind(this)}>
-                <TouchableOpacity
-                  style={style.locationButton}
-                  onPress={() => { this.getProfile(profile) }}>
-                  <View style={{ flex: 1, flexDirection: 'row', justifyContent: 'space-between' }}>
-                    <View>
-                      <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'flex-start', margin: 10 }}>
-                        <Image source={require('./assets/images/crosshairs.png')} style={{ width: 25, height: 25 }} />
-                        <Text style={style.h2}>{crop.profile_name}</Text>
-                      </View>
-                      <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'flex-start' }}>
-                        <Icon style={{ paddingLeft: 15 }} name='eye' color='black' size={18} />
-                        <Text style={style.t5}>{profile.objects.length} Sightings</Text>
-                      </View>
-                      <Text style={style.t5}>Last seen {Moment(new Date() - Moment(crop.photo_time)).format('DD')} days ago</Text>
-                    </View>
-                    <View style={{ width: thumbWidth, alignItems: 'center', justifyContent: 'center' }}>
-                      <Image source={{ uri: crop.path }} style={{ width: w, height: h }} />
-                    </View>
-                  </View>
-                </TouchableOpacity>
-              </SwipeRow>
-            );
-          })}
-        </ScrollView>
-      </SafeAreaView >
+        <View style={{ height: 40 }}>
+          <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'center' }}>
+            <Image source={require('./assets/images/crosshairs.png')} style={{ width: 25, height: 25 }} />
+            <Text style={style.t4}>{photos[0].profile_name}</Text>
+          </View>
+        </View>
+        <View style={{ height: screenHeight - 40 - headerHeight - footerHeight }}>
+          <Swiper showsButtons={true} loop={true} activeDotColor='rgb(255, 103, 0)' dotColor='#4E603E'>
+            <View>
+              <View style={{ alignItems: 'center' }}>
+                <View style={{ width: screenWidth, alignItems: 'center', justifyContent: 'center' }}>
+                  <Image source={{ uri: crop.path }} style={{ width: w, height: h }} />
+                </View>
+              </View>
+            </View>
+            <View>
+              <View>
+                <PhotoGrid photos={photos} showCrops={true} onRefresh={() => { this.props.navigation.goBack() }} />
+              </View>
+            </View>
+          </Swiper>
+        </View>
+      </SafeAreaView>
     );
   }
 
-  deleteProfile(profile, callback) {
-    console.log(profile);
-    Alert.alert(
-      'Delete Profile ' + profile.profile_name + '?', '', [
-      {
-        text: 'Yes',
-        onPress: () => {
-          this.db.deleteProfile(profile.profile_id).then(() => {
-            this.fetchData();
-          });
-          callback();
-        }
-      },
-      { text: 'No', onPress: callback }], { cancelable: false });
-  }
-
-  fetchData() {
-    this.db.selectProfiles().then((profiles) => {
-      this.setState({
-        isLoading: false,
-        profiles: profiles
-      });
-    }).catch((error) => {
-      console.log(error);
-    });
-  }
 }
